pyproject.toml
[project]
name = "dvd-rental-graphql"
version = "1.0.0"
description = "GraphQL API for DVD Rental System using Strawberry"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "strawberry-graphql[fastapi]>=0.250.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends.legacy:build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true
